b_dev/backend/requirements.txt
flask
flask-restful
python-dotenv
flask-sqlalchemy